reorganize fhircraft.fhir.resources.base into a module with split files for better maintanenance
