+++ fhircraft/fhir/resources/base/model.py
@@ -6,6 +6,7 @@ from itertools import zip_longest
 import operator
 import re
 import threading
+from unicodedata import name
 import warnings
 from typing import (
     Any,
@@ -460,6 +461,9 @@ class FHIRBaseModel(BaseModel, FHIRPathMixin):
         Args:
             value: The field value (can be FHIRBaseModel, list, or other)
         """
+
+        from fhircraft.fhir.resources.base import FHIRList, FHIRPrimitiveModel
+
         if isinstance(value, FHIRBaseModel):
             # Set _parent directly – no recursive descent needed.
             object.__setattr__(value, "_parent", self)
@@ -536,6 +540,9 @@ class FHIRBaseModel(BaseModel, FHIRPathMixin):
 
     def serialize_as_xml(self, name: str, **kwargs) -> xml.Element:
         """Serialize this instance as an XML element (not a string)."""
+
+        from fhircraft.fhir.resources.base import FHIRList
+
         element = xml.Element(f"{{{XML_NAMESPACE}}}{name}")
         for subelement_name in self.model_dump(**kwargs):
             if subelement := getattr(self, subelement_name, None):
@@ -563,6 +570,8 @@ class FHIRBaseModel(BaseModel, FHIRPathMixin):
     @staticmethod
     def _get_primitive_shadow_data(value: Any) -> "Any | None":
         """Return the _fieldname shadow dict/list for a primitive with id/extension, or None."""
+        from fhircraft.fhir.resources.base import FHIRPrimitiveModel
+
         if isinstance(value, FHIRPrimitiveModel):
             shadow: dict = {}
             if getattr(value, "id", None) is not None:
@@ -1021,456 +1030,6 @@ class FHIRBaseModel(BaseModel, FHIRPathMixin):
         return f"{type(self).__name__}({', '.join(self._get_repr_args())})"
 
 
-class FHIRPrimitiveModel(FHIRBaseModel):
-    """
-    Base class for FHIR primitive types.
-
-    FHIR primitives are represented as Pydantic models with a single `value` field that holds the actual primitive value.
-    This design allows us to attach extensions to primitive values while still treating them as simple types in most contexts.
-    """
-
-    _kind = "primitive-type"
-
-    value: Any | None = Field(default=None, description="The actual value")
-
-    @model_serializer
-    def serialize_root_value(self) -> Any:
-        if isinstance(self, FHIRPrimitiveModel):
-            return self.value
-        return self
-
-    def serialize_as_json(self, name: str) -> dict:
-        serialized = dict()
-        if self.value:
-            serialized[name] = self.value
-        if self.id or self.extension:
-            serialized[f"_{name}"] = {}
-            if self.id is not None:
-                serialized[f"_{name}"]["id"] = self.id
-            if self.extension is not None:
-                serialized[f"_{name}"]["extension"] = [
-                    ext.model_dump() for ext in self.extension
-                ]
-        return serialized
-
-    def serialize_as_xml(self, name: str, **kwargs) -> xml.Element:
-        attributes = {}
-        if self.value is not None:
-            attributes["value"] = (
-                str(self.value).lower()
-                if isinstance(self.value, bool)
-                else str(self.value)
-            )
-        if getattr(self, "id", None) is not None:
-            attributes["id"] = self.id
-        primitive = xml.Element(f"{{{XML_NAMESPACE}}}{name}", attrib=attributes)
-        for ext in self.extension or []:
-            primitive.append(ext.serialize_as_xml("extension", **kwargs))
-        return primitive
-
-    @classmethod
-    def parse_xml_to_dict(cls, element: xml.Element) -> dict:
-        element_name = element.tag.split("}", 1)[-1]  # Remove namespace
-        # Collect extensions if present using the correct Extension type from the field
-        extensions = []
-        extension_field = cls.model_fields.get("extension")
-        if extension_field is not None:
-            extension_type = next(get_all_models_from_field(extension_field), None)
-            if extension_type is not None and issubclass(extension_type, FHIRBaseModel):
-                for ext_element in element.findall(f"{{{XML_NAMESPACE}}}extension"):
-                    ext_result = extension_type.parse_xml_to_dict(ext_element)
-                    ext_data = ext_result.get("extension")
-                    if ext_data is not None:
-                        extensions.append(ext_data)
-        # Extract the "id" attribute if present
-        elem_id = element.attrib.get("id")
-        # Extract the "value" attribute if present, converting boolean strings to actual booleans
-        value = element.attrib.get("value")
-        if value is not None:
-            # Convert boolean strings
-            if value == "true":
-                value = True
-            elif value == "false":
-                value = False
-
-        deserialized_data = {}
-        if value is not None:
-            deserialized_data[element_name] = value
-        if elem_id is not None or extensions:
-            deserialized_data[f"_{element_name}"] = {}
-        if elem_id is not None:
-            deserialized_data[f"_{element_name}"]["id"] = elem_id
-        if extensions:
-            deserialized_data[f"_{element_name}"]["extension"] = extensions
-        return deserialized_data
-
-    @model_validator(mode="before")
-    @classmethod
-    def coerce_root_value(cls, data: Any) -> Any:
-        if isinstance(data, FHIRPrimitiveModel):
-            return {"value": data.value}
-        elif data is not None and not isinstance(data, (dict, BaseModel)):
-            return {"value": data}
-        return data
-
-    def __init__(self, __value: Any | None = None, **data):
-        if __value is not None:
-            data.setdefault("value", __value)
-        super().__init__(**data)
-
-    def __getattr__(self, name: str):
-        v = self.value
-        if v is None:
-            raise AttributeError(f"Cannot access '{name}' because value is None")
-        return getattr(v, name)
-
-    def __getitem__(self, key):
-        v = self.value
-        if v is None:
-            raise TypeError("Cannot index into None value")
-        return v[key]
-
-    def __eq__(self, other: Any) -> bool:
-        if isinstance(other, FHIRPrimitiveModel):
-            return self.value == other.value
-        return self.value == other
-
-    def __gt__(self, other):
-        return (
-            self.value > other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value > other
-        )
-
-    def __rgt__(self, other):
-        return (
-            self.value < other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value < other
-        )
-
-    def __lt__(self, other):
-        return (
-            self.value < other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value < other
-        )
-
-    def __rlt__(self, other):
-        return (
-            self.value > other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value > other
-        )
-
-    def __ge__(self, other):
-        return (
-            self.value >= other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value >= other
-        )
-
-    def __rge__(self, other):
-        return (
-            self.value <= other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value <= other
-        )
-
-    def __le__(self, other):
-        return (
-            self.value <= other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value <= other
-        )
-
-    def __rle__(self, other):
-        return (
-            self.value >= other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value >= other
-        )
-
-    def __add__(self, other):
-        return (
-            self.value + other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value + other
-        )
-
-    def __radd__(self, other):
-        return (
-            other.value + self.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else other + self.value
-        )
-
-    def __sub__(self, other):
-        return (
-            self.value - other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value - other
-        )
-
-    def __rsub__(self, other):
-        return (
-            other.value - self.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else other - self.value
-        )
-
-    def __mul__(self, other):
-        return (
-            self.value * other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value * other
-        )
-
-    def __rmul__(self, other):
-        return (
-            self.value * other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value * other
-        )
-
-    def __truediv__(self, other):
-        return (
-            self.value / other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value / other
-        )
-
-    def __rtruediv__(self, other):
-        return (
-            other.value / self.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else other / self.value
-        )
-
-    def __mod__(self, other):
-        return (
-            self.value % other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value % other
-        )
-
-    def __rmod__(self, other):
-        return (
-            other.value % self.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else other % self.value
-        )
-
-    def __floordiv__(self, other):
-        return (
-            self.value // other.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else self.value // other
-        )
-
-    def __rfloordiv__(self, other):
-        return (
-            other.value // self.value  # type: ignore
-            if isinstance(other, FHIRPrimitiveModel)
-            else other // self.value
-        )
-
-    def __neg__(self):
-        return -self.value  # type: ignore
-
-    def __pos__(self):
-        return +self.value  # type: ignore
-
-    def __abs__(self):
-        return abs(self.value)  # type: ignore
-
-    def __hash__(self) -> int:
-        return hash(self.value)
-
-    def __str__(self) -> str:
-        return str(self.value) if self.value is not None else ""
-
-    def __repr__(self) -> str:
-        return repr(self.value)
-
-
-class StringBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR string types
-    """
-
-    _kind = "string"
-    value: str | None = None
-
-
-class BooleanBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR boolean types
-    """
-
-    _kind = "boolean"
-    value: bool | None = None
-
-
-class DecimalBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR decimal types
-    """
-
-    _kind = "decimal"
-    value: float | None = None
-
-
-class DateBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR date types
-    """
-
-    _kind = "date"
-    value: str | None = None
-
-
-class DateTimeBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR dateTime types
-    """
-
-    _kind = "dateTime"
-    value: str | None = None
-
-
-class TimeBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR time types
-    """
-
-    _kind = "time"
-    value: str | None = None
-
-
-class InstantBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR instant types
-    """
-
-    _kind = "instant"
-    value: str | None = None
-
-
-class Base64BinaryBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR base64Binary types
-    """
-
-    _kind = "base64Binary"
-    value: str | None = None
-
-
-class UriBase(StringBase):
-    """
-    A release-independent base metaclass for FHIR uri types
-    """
-
-    _kind = "uri"
-
-
-class UrlBase(UriBase):
-    """
-    A release-independent base metaclass for FHIR url types
-    """
-
-    _kind = "url"
-
-
-class CanonicalBase(UriBase):
-    """
-    A release-independent base metaclass for FHIR canonical types
-    """
-
-    _kind = "canonical"
-
-
-class OidBase(UriBase):
-    """
-    A release-independent base metaclass for FHIR oid types
-    """
-
-    _kind = "oid"
-
-
-class UuidBase(UriBase):
-    """
-    A release-independent base metaclass for FHIR uuid types
-    """
-
-    _kind = "uuid"
-
-
-class CodeBase(StringBase):
-    """
-    A release-independent base metaclass for FHIR code types
-    """
-
-    _kind = "code"
-
-
-class IdBase(StringBase):
-    """
-    A release-independent base metaclass for FHIR id types
-    """
-
-    _kind = "id"
-
-
-class MarkdownBase(StringBase):
-    """
-    A release-independent base metaclass for FHIR markdown types
-    """
-
-    _kind = "markdown"
-
-
-class XhtmlBase(StringBase):
-    """
-    A release-independent base metaclass for FHIR xhtml types
-    """
-
-    _kind = "xhtml"
-
-
-class IntegerBase(FHIRPrimitiveModel):
-    """
-    A release-independent base metaclass for FHIR integer types
-    """
-
-    _kind = "integer"
-    value: int | None = None
-
-
-class PositiveIntBase(IntegerBase):
-    """
-    A release-independent base metaclass for FHIR positiveInt types
-    """
-
-    _kind = "positiveInt"
-
-
-class UnsignedIntBase(IntegerBase):
-    """
-    A release-independent base metaclass for FHIR unsignedInt types
-    """
-
-    _kind = "unsignedInt"
-
-
-class Integer64Base(IntegerBase):
-    """
-    A release-independent base metaclass for FHIR integer64 types (R5)
-    """
-
-    _kind = "integer64"
-
-
 class FHIRSliceModel(FHIRBaseModel):
     """
     Base class for representation of FHIR profiled slices as Pydantic objects.
@@ -1501,73 +1060,3 @@ class FHIRSliceModel(FHIRBaseModel):
         Returns `True` if the instance has been modified, `False` otherwise.
         """
         return self != self.__class__.model_construct_with_slices()
-
-
-class FHIRList(list):
-    """
-    Custom list wrapper that maintains parent context on mutations.
-
-    This list automatically propagates _parent, _root_resource, _resource, and _index context
-    to FHIRBaseModel items when they are added via append, extend, insert, or __setitem__.
-    """
-
-    def __init__(self, items=None, parent=None, root=None, resource=None):
-        """Initialize FHIRList with items and context.
-
-        ``root`` and ``resource`` are accepted for backwards-compatibility but
-        are no longer stored; they are resolved lazily via ``_parent`` on items.
-        """
-        super().__init__(items or [])
-        self._parent = parent
-        self._propagate_context()
-
-    def _propagate_context(self):
-        """Set _parent and _index on all current FHIRBaseModel items."""
-        if self._parent is None:
-            return
-
-        for index, item in enumerate(self):
-            if isinstance(item, FHIRBaseModel):
-                object.__setattr__(item, "_parent", self._parent)
-                object.__setattr__(item, "_index", index)
-
-    def append(self, item):
-        """Append item and propagate context."""
-        super().append(item)
-        if isinstance(item, FHIRBaseModel):
-            object.__setattr__(item, "_parent", self._parent)
-            object.__setattr__(item, "_index", len(self) - 1)
-
-    def extend(self, items):
-        """Extend list and propagate context to new items."""
-        start_index = len(self)
-        super().extend(items)
-        for offset, item in enumerate(items):
-            if isinstance(item, FHIRBaseModel):
-                object.__setattr__(item, "_parent", self._parent)
-                object.__setattr__(item, "_index", start_index + offset)
-
-    def insert(self, index, item):
-        """Insert item and propagate context."""
-        super().insert(index, item)
-        if isinstance(item, FHIRBaseModel):
-            object.__setattr__(item, "_parent", self._parent)
-            object.__setattr__(item, "_index", index)
-        # Re-index all items after insertion point
-        for i in range(index + 1, len(self)):
-            if isinstance(self[i], FHIRBaseModel):
-                object.__setattr__(self[i], "_index", i)
-
-    def __setitem__(self, index, item):
-        """Set item and propagate context."""
-        super().__setitem__(index, item)
-        if isinstance(item, FHIRBaseModel):
-            if isinstance(index, int):
-                object.__setattr__(item, "_parent", self._parent)
-                object.__setattr__(item, "_index", index)
-            else:
-                # Slice assignment – re-propagate to fix indices.
-                self._propagate_context()
-        elif isinstance(item, list):
-            # Slice assignment with a plain list – re-propagate to fix indices.
-            self._propagate_context()
